+++ Package.swift
@@ -3,7 +3,7 @@
 import PackageDescription
 
 let sdkVersion = Version("1.1.1")
-let coreVersion = Version("14.5.1")
+let coreVersion = Version("14.7.0")
 
 var cxxSettings: [CXXSetting] = [
     .define("REALM_ENABLE_SYNC", to: "1"),
@@ -68,13 +68,13 @@ let package = Package(
             targets: ["realm-cpp-sdk"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/realm/realm-core.git", revision: "316889b967f845fbc10b4422f96c7eadd47136f2")
+        .package(url: "https://github.com/realm/realm-core.git", exact: coreVersion)
     ],
     targets: [
         cppSdkTarget,
         .target(
             name: "Catch2Generated",
-            path: "realm-core/external/generated",
+            path: "tests/external/generated",
             // this file was manually generated with catch v3.0.1
             // and should be regenerated when catch is upgraded
             resources: [.copy("catch2/catch_user_config.hpp")],